Framework: Make build work with Docker with a smaller build file size
