+++ client/my-sites/upgrades/index.js
@@ -28,7 +28,7 @@ function getCommonHandlers( { noSitePath = paths.domainManagementRoot(), warnIfJ
 	}
 
 	if ( warnIfJetpack ) {
-		handlers.push( controller.jetPackWarning );
+		handlers.push( upgradesController.jetPackWarning );
 	}
 
 	return handlers;
@@ -155,7 +155,7 @@ module.exports = function() {
 			controller.siteSelection,
 			upgradesController.domainsAddHeader,
 			upgradesController.redirectToAddMappingIfVipSite(),
-			controller.jetPackWarning,
+			upgradesController.jetPackWarning,
 			controller.sites
 		);
 
@@ -163,7 +163,7 @@ module.exports = function() {
 			'/domains/add/mapping',
 			controller.siteSelection,
 			upgradesController.domainsAddHeader,
-			controller.jetPackWarning,
+			upgradesController.jetPackWarning,
 			controller.sites
 		);
 
@@ -171,7 +171,7 @@ module.exports = function() {
 			'/domains/add/site-redirect',
 			controller.siteSelection,
 			upgradesController.domainsAddRedirectHeader,
-			controller.jetPackWarning,
+			upgradesController.jetPackWarning,
 			controller.sites
 		);
 
@@ -180,7 +180,7 @@ module.exports = function() {
 			controller.navigation,
 			upgradesController.redirectIfNoSite( '/domains/add' ),
 			upgradesController.redirectToAddMappingIfVipSite(),
-			controller.jetPackWarning,
+			upgradesController.jetPackWarning,
 			upgradesController.domainSearch
 		);
 
@@ -189,7 +189,7 @@ module.exports = function() {
 			controller.navigation,
 			upgradesController.redirectIfNoSite( '/domains/add' ),
 			upgradesController.redirectToAddMappingIfVipSite(),
-			controller.jetPackWarning,
+			upgradesController.jetPackWarning,
 			upgradesController.domainSearch
 		);
 
@@ -197,7 +197,7 @@ module.exports = function() {
 			controller.siteSelection,
 			controller.navigation,
 			upgradesController.redirectIfNoSite( '/domains/add' ),
-			controller.jetPackWarning,
+			upgradesController.jetPackWarning,
 			upgradesController.googleAppsWithRegistration
 		);
 
@@ -205,7 +205,7 @@ module.exports = function() {
 			controller.siteSelection,
 			controller.navigation,
 			upgradesController.redirectIfNoSite( '/domains/add/mapping' ),
-			controller.jetPackWarning,
+			upgradesController.jetPackWarning,
 			upgradesController.mapDomain
 		);
 
@@ -213,7 +213,7 @@ module.exports = function() {
 			controller.siteSelection,
 			controller.navigation,
 			upgradesController.redirectIfNoSite( '/domains/add/site-redirect' ),
-			controller.jetPackWarning,
+			upgradesController.jetPackWarning,
 			upgradesController.siteRedirect
 		);
 	}
@@ -228,7 +228,7 @@ module.exports = function() {
 		'/domains/:site',
 		controller.siteSelection,
 		controller.navigation,
-		controller.jetPackWarning,
+		upgradesController.jetPackWarning,
 		domainManagementController.domainManagementIndex
 	);
